requirements.txt
astunparse
blendmodes
accelerate
basicsr
fonts
font-roboto
gfpgan
gradio==3.23
invisible-watermark
numpy
omegaconf
opencv-contrib-python
requests
piexif
Pillow
pytorch_lightning==1.7.7
realesrgan
scikit-image>=0.19
timm==0.4.12
transformers==4.25.1
torch
einops
jsonmerge
clean-fid
resize-right
torchdiffeq
kornia
lark
svglib
k_diffusion==0.0.14
inflection
GitPython
torchsde
safetensors
psutil
rich
redis
pymysql
aiofiles
aiohttp
altair>=4.2.2
fastapi
ffmpy
gradio_client>=0.0.4
httpx
huggingface_hub>=0.13.0
Jinja2
markdown-it-py[linkify]>=2.0.0
mdit-py-plugins<=0.3.3
markupsafe
matplotlib
numpy
orjson
pandas
pillow
pydantic
python-multipart
pydub
pyyaml
requests
semantic_version
typing_extensions
uvicorn
websockets>=10.0
nvidia-ml-py==11.525.84
loguru
pycryptodome
python-redis-lock==4.0.0
pymongo==4.3.3
s3fs
pygments
oss2
tomesd
